test(craft): share sandbox_proxy flow/sandbox/resolver test factories
